pyproject.toml
[project]
name = "seat-sense"
version = "0.1.0"
description = "Realtime Auditorium Seat Occuapncy Detection & Attendence System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.13.3",
    "asyncpg>=0.29.0",
    "bcrypt>=4.2.0",
    "fastapi[standard]>=0.115.0",
    "greenlet>=3.1.1",
    "jwt>=1.3.1",
    "opencv-python-headless>=4.10.0.84",
    "psycopg2>=2.9.9",
    "pydantic-settings>=2.5.2",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "scikit-image>=0.24.0",
    "sqlalchemy>=2.0.35",
    "websockets>=13.1",
]
